+++ internal/web/ui.go
@@ -4,3 +4,6 @@ import _ "embed"
 
 //go:embed static/index.html
 var indexHTML string
+
+//go:embed static/favicon.png
+var faviconPNG []byte